requirements.txt
transformers
huggingface-hub
faiss-cpu
numpy
pandas
streamlit
requests
beautifulsoup4
pdfplumber
groq